+++ starknet_py/net/client_test.py
@@ -10,11 +10,11 @@ from starknet_py.net.client_models import (
     Transaction,
     TransactionV3,
 )
-from starknet_py.net.client_utils import (
-    _create_broadcasted_txn,
+from starknet_py.net.client_utils import _create_broadcasted_txn
+from starknet_py.net.full_node_client import (
     _create_broadcasted_txn_prev,
+    _to_storage_key,
 )
-from starknet_py.net.full_node_client import _to_storage_key
 from starknet_py.net.http_client import RpcHttpClient, ServerError
 from starknet_py.tests.e2e.fixtures.constants import MAX_FEE, MAX_RESOURCE_BOUNDS_L1
